+++ lib/pdftk_forms.rb
@@ -1,5 +1,7 @@
 require 'builder'
+require 'pdftk_forms/call'
 require 'pdftk_forms/field'
 require 'pdftk_forms/fdf'
 require 'pdftk_forms/xfdf'
 require 'pdftk_forms/wrapper'
+
