+++ spec/pdftk_forms/call_spec.rb
@@ -1,0 +1,11 @@
+require 'spec_helper'
+
+describe PdftkForms::Call do
+  context "new" do
+    it "should set the default path" do
+    end
+    it "should allow for custom paths" do
+    end
+  end
+end
+
